requirements.txt
fairscale~=0.4.13
torch~=2.0.1
sentencepiece~=0.1.99
pathlib~=1.0.1
llama~=0.0.1
setuptools~=68.0.0
fire~=0.5.0
openai~=0.28.1
python-dotenv~=1.0.0
pypdf2~=3.0.1
langchain~=0.0.302
requests~=2.31.0
markdown~=3.4.1
ipython~=8.12.0
numpy~=1.24.3
matplotlib~=3.7.1
markdown2~=2.4.10
future~=0.18.3
django~=4.2.6
pypdfium2~=4.20.0
fastapi~=0.103.2
pillow~=9.4.0
tqdm~=4.65.0
pypdf~=3.16.2
orjson~=3.9.7
transformers~=4.34.0
unidecode~=1.2.0
beautifulsoup4~=4.12.2
soupsieve~=2.4
levenshtein~=0.22.0
scikit-learn~=1.3.0
opencv-python~=4.8.0.76
timm~=0.5.4
torchvision~=0.15.2
nltk~=3.8.1
albumentations~=1.3.1
lightning~=2.0.9.post0
sconf~=0.2.5
pywebio~=1.8.3
katex~=0.0.4